Retire the service review dashboard (P3.4)
The dashboard's two jobs now live elsewhere: triage in the review inbox,
editing inline on the service workbench. Its URL 302s to the inbox so
bookmarks keep working; ServiceReviewDashboardQuery stays (the inbox,
workbench, and attention counts consume it).

Inbox section rows deep-link Edit to the workbench #section-{id} anchor
when the run matched a service, and hide the link otherwise (the inline
approve/reject/requeue quick actions remain). Remaining dashboard tests
migrate to ShowChurchServiceTest: review-reason chips (including
inferred-vs-unmatched song labels), edit-buffer seeding on refresh,
approve/reject/requeue, save validation, stale-section guard, the four
merge rejection cases, and children's-talk speaker edits. The empty-state
and service-only-group behaviours were already pinned by ReviewInboxTest.
