+++ tests/Feature/Livewire/Admin/ChurchServices/ShowChurchServiceTest.php
@@ -9,6 +9,7 @@ use App\Enums\ProcessingStatus;
 use App\Enums\SermonService;
 use App\Enums\ServiceSectionPublicationStatus;
 use App\Enums\ServiceSectionType;
+use App\Jobs\PublishApprovedServiceSection;
 use App\Livewire\Admin\ChurchServices\ListSectionPublications;
 use App\Livewire\Admin\ChurchServices\ProcessingReviewList;
 use App\Livewire\Admin\ChurchServices\ShowChurchService;
@@ -18,6 +19,7 @@ use App\Models\ChurchService;
 use App\Models\ChurchServiceItem;
 use App\Models\LivestreamSegment;
 use App\Models\MediaProcessingLog;
+use App\Models\Preacher;
 use App\Models\Sermon;
 use App\Models\ServiceSection;
 use App\Models\User;
@@ -609,6 +611,391 @@ class ShowChurchServiceTest extends TestCase
             ->test(ShowChurchService::class, ['churchService' => $service])
             ->call('confirmRunSegment', $run->id, $segment->id)
             ->assertForbidden();
+    }
+
+    // -------------------------------------------------------------------------
+    // Migrated from AdminServiceReviewDashboardTest (P3.4) — the dashboard is
+    // retired; these behaviours now live on the workbench.
+    // -------------------------------------------------------------------------
+
+    #[Test]
+    public function it_renders_review_reason_chips_distinguishing_song_match_types(): void
+    {
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $songItem = ChurchServiceItem::factory()->create([
+            'church_service_id' => $service->id,
+            'position' => 1,
+            'type' => 'songs',
+            'title' => 'Opening Song',
+            'song_id' => null,
+        ]);
+
+        ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'church_service_item_id' => $songItem->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 1,
+            'title' => 'Unknown Song',
+            'needs_manual_review' => true,
+            'confidence' => 0.72,
+            'song_match_type' => 'unmatched',
+            'metadata' => [
+                'review_reason' => 'expected_type_mismatch',
+                'oos_alignment' => ['song_match_type' => 'unmatched'],
+            ],
+            'publication_status' => ServiceSectionPublicationStatus::PendingApproval->value,
+        ]);
+
+        ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 2,
+            'title' => 'Inferred Song',
+            'needs_manual_review' => true,
+            'confidence' => 0.72,
+            'song_match_type' => 'inferred',
+            'metadata' => [
+                'review_reason' => 'song_alignment_inferred',
+                'oos_alignment' => ['song_match_type' => 'inferred'],
+            ],
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->assertSee('Unknown Song')
+            ->assertSee('Manual review')
+            ->assertSee('Pending approval')
+            ->assertSee('Low confidence')
+            ->assertSee('Unmatched song')
+            ->assertSee('Inferred song label');
+    }
+
+    #[Test]
+    public function it_does_not_seed_edit_buffers_for_review_rows_created_after_mount(): void
+    {
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $existingSection = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Welcome->value,
+            'section_order' => 1,
+            'title' => 'Existing review row',
+            'needs_manual_review' => true,
+        ]);
+
+        $component = Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->assertSee('Existing review row');
+
+        $this->assertArrayHasKey($existingSection->id, $component->get('sectionEdits'));
+
+        $newSection = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Prayer->value,
+            'section_order' => 2,
+            'title' => 'New review row',
+            'needs_manual_review' => true,
+        ]);
+
+        $component
+            ->call('$refresh')
+            ->assertSee('New review row');
+
+        $this->assertArrayNotHasKey($newSection->id, $component->get('sectionEdits'));
+    }
+
+    #[Test]
+    public function approve_and_reject_actions_update_publication_status_on_the_workbench(): void
+    {
+        Queue::fake();
+
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $approveSection = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Welcome->value,
+            'section_order' => 1,
+            'publication_status' => ServiceSectionPublicationStatus::PendingApproval->value,
+            'extracted_video_path' => 'sermons/sections/'.$run->id.'/video.mp4',
+            'extracted_audio_path' => 'sermons/audio/section-'.$run->id.'.mp3',
+        ]);
+
+        $rejectSection = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Welcome->value,
+            'section_order' => 2,
+            'publication_status' => ServiceSectionPublicationStatus::PendingApproval->value,
+        ]);
+
+        Storage::disk('public')->put('sermons/sections/'.$run->id.'/video.mp4', 'video');
+        Storage::disk('public')->put('sermons/audio/section-'.$run->id.'.mp3', 'audio');
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->call('approve', $approveSection->id)
+            ->assertDispatched('notify', type: 'success', message: 'Section approved and publish job queued.')
+            ->call('reject', $rejectSection->id)
+            ->assertDispatched('notify', type: 'success', message: 'Section rejected.');
+
+        $this->assertSame(ServiceSectionPublicationStatus::Approved, $approveSection->fresh()->publication_status);
+        $this->assertSame(ServiceSectionPublicationStatus::Rejected, $rejectSection->fresh()->publication_status);
+        Queue::assertPushed(PublishApprovedServiceSection::class);
+    }
+
+    #[Test]
+    public function requeue_moves_a_rejected_section_back_to_pending_on_the_workbench(): void
+    {
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $section = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'publication_status' => ServiceSectionPublicationStatus::Rejected->value,
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->call('requeue', $section->id)
+            ->assertDispatched('notify', type: 'success', message: 'Section moved back to pending approval.');
+
+        $this->assertSame(ServiceSectionPublicationStatus::PendingApproval, $section->fresh()->publication_status);
+    }
+
+    #[Test]
+    public function save_section_requires_valid_type_and_title(): void
+    {
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $section = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'needs_manual_review' => true,
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->set("sectionEdits.{$section->id}.section_type", 'invalid')
+            ->set("sectionEdits.{$section->id}.title", '')
+            ->call('saveSection', $section->id)
+            ->assertHasErrors(['section_type', 'title']);
+    }
+
+    #[Test]
+    public function it_blocks_saving_a_section_that_is_no_longer_in_the_review_queue(): void
+    {
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $section = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'needs_manual_review' => false,
+            'confidence' => 0.98,
+            'publication_status' => ServiceSectionPublicationStatus::NotApplicable->value,
+            'section_type' => ServiceSectionType::Welcome->value,
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->set("sectionEdits.{$section->id}.section_type", ServiceSectionType::Prayer->value)
+            ->set("sectionEdits.{$section->id}.title", 'Updated Title')
+            ->call('saveSection', $section->id)
+            ->assertDispatched('notify', type: 'error', message: 'Section is no longer awaiting review.');
+
+        $section->refresh();
+        $this->assertSame(ServiceSectionType::Welcome, $section->section_type);
+        $this->assertNotSame('Updated Title', $section->title);
+    }
+
+    #[Test]
+    public function it_rejects_merging_sections_from_different_processing_runs(): void
+    {
+        Queue::fake();
+
+        [$service, $runA] = $this->workbenchServiceWithRun();
+
+        $runB = MediaProcessingLog::factory()->livestream()->completed()->create([
+            'extracted_date' => '2026-06-07',
+            'extracted_service' => SermonService::Morning->value,
+            'sermon_id' => null,
+        ]);
+
+        $first = ServiceSection::factory()->create([
+            'media_processing_log_id' => $runA->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 1,
+            'needs_manual_review' => true,
+            'start_time' => 100.0,
+            'end_time' => 107.0,
+        ]);
+
+        $second = ServiceSection::factory()->create([
+            'media_processing_log_id' => $runB->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 1,
+            'needs_manual_review' => true,
+            'start_time' => 107.0,
+            'end_time' => 231.0,
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->call('initiateMerge', $first->id, $second->id)
+            ->call('confirmMerge')
+            ->assertDispatched('notify', type: 'error', message: 'Both sections must belong to the same processing run.');
+
+        $this->assertDatabaseHas('service_sections', ['id' => $first->id]);
+        $this->assertDatabaseHas('service_sections', ['id' => $second->id]);
+    }
+
+    #[Test]
+    public function it_rejects_merging_sections_of_different_types(): void
+    {
+        Queue::fake();
+
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $first = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 1,
+            'needs_manual_review' => true,
+            'start_time' => 100.0,
+            'end_time' => 107.0,
+        ]);
+
+        $second = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::BibleReading->value,
+            'section_order' => 2,
+            'needs_manual_review' => true,
+            'start_time' => 107.0,
+            'end_time' => 231.0,
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->call('initiateMerge', $first->id, $second->id)
+            ->call('confirmMerge')
+            ->assertDispatched('notify', type: 'error', message: 'Both sections must have the same section type.');
+    }
+
+    #[Test]
+    public function it_rejects_merging_a_published_section(): void
+    {
+        Queue::fake();
+
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $first = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 1,
+            'needs_manual_review' => true,
+            'start_time' => 100.0,
+            'end_time' => 107.0,
+            'publication_status' => ServiceSectionPublicationStatus::Published->value,
+        ]);
+
+        $second = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 2,
+            'needs_manual_review' => true,
+            'start_time' => 107.0,
+            'end_time' => 231.0,
+            'publication_status' => ServiceSectionPublicationStatus::NotApplicable->value,
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->call('initiateMerge', $first->id, $second->id)
+            ->call('confirmMerge')
+            ->assertDispatched('notify', type: 'error', message: 'Published sections cannot be merged.');
+    }
+
+    #[Test]
+    public function it_rejects_merging_when_another_section_sits_between_the_candidates(): void
+    {
+        Queue::fake();
+
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $first = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 5,
+            'needs_manual_review' => true,
+            'start_time' => 100.0,
+            'end_time' => 107.0,
+        ]);
+
+        // Un-flagged section at order 6 (between the merge candidates by section_order)
+        ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Prayer->value,
+            'section_order' => 6,
+            'needs_manual_review' => false,
+            'start_time' => 300.0,
+            'end_time' => 400.0,
+        ]);
+
+        // Starts close to $first (gap 1.5s ≤ max 2s) so the gap check passes and
+        // the between-sections check triggers instead.
+        $third = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::Song->value,
+            'section_order' => 7,
+            'needs_manual_review' => true,
+            'start_time' => 108.5,
+            'end_time' => 244.0,
+        ]);
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->call('initiateMerge', $first->id, $third->id)
+            ->call('confirmMerge')
+            ->assertDispatched('notify', type: 'error', message: 'There are other sections between these two — they cannot be merged.');
+
+        $this->assertDatabaseHas('service_sections', ['id' => $first->id]);
+        $this->assertDatabaseHas('service_sections', ['id' => $third->id]);
+    }
+
+    #[Test]
+    public function it_saves_speaker_edits_for_childrens_talk_sections(): void
+    {
+        [$service, $run] = $this->workbenchServiceWithRun();
+
+        $preacher = Preacher::factory()->create(['name' => 'Rev Override']);
+
+        $section = ServiceSection::factory()->create([
+            'media_processing_log_id' => $run->id,
+            'section_type' => ServiceSectionType::ChildrensTalk->value,
+            'title' => "Children's Talk",
+            'needs_manual_review' => true,
+            'publication_status' => ServiceSectionPublicationStatus::NotApplicable->value,
+            'extracted_video_path' => 'sermons/sections/speaker-test/video.mp4',
+            'extracted_audio_path' => 'sermons/audio/speaker-test.mp3',
+            'metadata' => [
+                'review_reason' => 'childrens_talk_speaker_ambiguous',
+                'childrens_talk_speaker' => [
+                    'predicted' => ['outcome' => 'ambiguous', 'preacher_name' => 'Unknown', 'confidence' => 0.5, 'reason' => 'Too close to call.'],
+                ],
+            ],
+        ]);
+
+        Storage::disk('public')->put('sermons/sections/speaker-test/video.mp4', 'video');
+        Storage::disk('public')->put('sermons/audio/speaker-test.mp3', 'audio');
+
+        Livewire::actingAs($this->admin)
+            ->test(ShowChurchService::class, ['churchService' => $service])
+            ->set("speakerEdits.{$section->id}.preacher_id", (string) $preacher->id)
+            ->set("speakerEdits.{$section->id}.speaker_name", '')
+            ->call('saveSection', $section->id)
+            ->assertDispatched('notify', type: 'success', message: 'Section changes saved.');
+
+        $section->refresh();
+
+        $this->assertFalse($section->needs_manual_review);
+        $this->assertSame(ServiceSectionPublicationStatus::PendingApproval, $section->publication_status);
+        $this->assertSame('Rev Override', $section->metadata['childrens_talk_speaker']['reviewed']['preacher_name'] ?? null);
     }
 
     #[Test]
